[#520] refactor: 어드민 API 클라이언트를 dashboard-api로 교체
- 기존 auth-api(app-session 쿠키)에서 dashboard-api(dashboard-session 쿠키)로 교체
- 대상: getAdminClubs, getClubApplications, getClubMasterApplications,
         approveClubApplication, approveClubMasterApplication,
         rejectClubApplication, rejectClubMasterApplication
